[RFC] Mixin Refactor
