+++ l10n_br_repair/models/__init__.py
@@ -1,7 +1,7 @@
 from . import res_company
 from . import res_config_settings
 from . import repair_order
-from . import repair_line_mixin_methods
-from . import repair_line_mixin
+from . import fiscal_line_mixin
 from . import repair_line
 from . import repair_fee
+
